+++ tests/organigram/test_helpers.py
@@ -1,28 +1,28 @@
 from mex.common.models import ExtractedOrganizationalUnit
 from mex.common.organigram.helpers import (
-    build_child_map,
-    find_all_descendants,
-    find_first_level_descendants,
+    build_child_unit_map,
+    get_all_descendant_unit_ids,
     get_extracted_organizational_unit_with_parents,
+    get_first_level_child_unit_ids,
 )
 from mex.common.organigram.models import OrganigramUnit
 from mex.common.types import MergedOrganizationIdentifier, MergedPrimarySourceIdentifier
 
 
-def test_build_child_map(
+def test_build_child_unit_map(
     child_unit: OrganigramUnit, parent_unit: OrganigramUnit
 ) -> None:
-    test_map = build_child_map(
+    test_map = build_child_unit_map(
         [child_unit, child_unit, parent_unit, child_unit, parent_unit]
     )
 
     assert test_map == {"parent-unit": ["child-unit", "child-unit", "child-unit"]}
 
 
-def test_find_all_descendants(
+def test_get_all_descendant_unit_ids(
     child_unit: OrganigramUnit, parent_unit: OrganigramUnit
 ) -> None:
-    child_ids = find_all_descendants(
+    child_ids = get_all_descendant_unit_ids(
         [child_unit, child_unit, child_unit, parent_unit, parent_unit],
         str(parent_unit.identifier),
     )
@@ -30,10 +30,10 @@ def test_find_all_descendants(
     assert child_ids == [child_unit.identifier]
 
 
-def test_find_first_level_descendants(
+def test_get_first_level_child_unit_ids(
     child_unit: OrganigramUnit, parent_unit: OrganigramUnit
 ) -> None:
-    child_ids = find_first_level_descendants(
+    child_ids = get_first_level_child_unit_ids(
         [child_unit, child_unit, child_unit, parent_unit, parent_unit],
         str(parent_unit.identifier),
     )