+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- new function `get_first_level_child_unit_ids` to find direct child units of parent units
 - set up integration tests against slapd using docker compose
 - added `merge_rules` to merge two rules of the same type
 - added `merge_rule_set_responses` to merge two rule set responses